+++ rotkehlchen/chain/ethereum/modules/curve/decoder.py
@@ -4,10 +4,10 @@ from typing import TYPE_CHECKING, Any, Callable, Dict, List, Mapping, Optional, 
 from rotkehlchen.accounting.structures.base import HistoryBaseEntry
 from rotkehlchen.accounting.structures.types import HistoryEventSubType, HistoryEventType
 from rotkehlchen.assets.asset import EvmToken
-from rotkehlchen.chain.ethereum.decoding.interfaces import DecoderInterface
-from rotkehlchen.chain.ethereum.decoding.structures import ActionItem
-from rotkehlchen.chain.ethereum.structures import EthereumTxReceiptLog
 from rotkehlchen.chain.ethereum.types import string_to_evm_address
+from rotkehlchen.chain.evm.decoding.interfaces import DecoderInterface
+from rotkehlchen.chain.evm.decoding.structures import ActionItem
+from rotkehlchen.chain.evm.structures import EvmTxReceiptLog
 from rotkehlchen.constants.assets import A_ETH
 from rotkehlchen.errors.asset import UnknownAsset, WrongAssetType
 from rotkehlchen.logging import RotkehlchenLogsAdapter
@@ -18,8 +18,8 @@ from .constants import CPT_CURVE
 from .pools_cache import read_curve_pools
 
 if TYPE_CHECKING:
-    from rotkehlchen.chain.ethereum.decoding.base import BaseDecoderTools
     from rotkehlchen.chain.ethereum.manager import EthereumManager
+    from rotkehlchen.chain.evm.decoding.base import BaseDecoderTools
     from rotkehlchen.user_messages import MessagesAggregator
 
 
@@ -54,7 +54,7 @@ class CurveDecoder(DecoderInterface):  # lgtm[py/missing-call-to-init]
 
     def _decode_curve_remove_events(
         self,
-        tx_log: EthereumTxReceiptLog,
+        tx_log: EvmTxReceiptLog,
         transaction: EvmTransaction,
         decoded_events: List[HistoryBaseEntry],
         user_address: ChecksumEvmAddress,
@@ -105,7 +105,7 @@ class CurveDecoder(DecoderInterface):  # lgtm[py/missing-call-to-init]
 
     def _decode_curve_deposit_events(
         self,
-        tx_log: EthereumTxReceiptLog,
+        tx_log: EvmTxReceiptLog,
         decoded_events: List[HistoryBaseEntry],
         user_address: ChecksumEvmAddress,
     ) -> Tuple[Optional[HistoryBaseEntry], List[ActionItem]]:
@@ -171,10 +171,10 @@ class CurveDecoder(DecoderInterface):  # lgtm[py/missing-call-to-init]
 
     def _decode_curve_events(  # pylint: disable=no-self-use
         self,
-        tx_log: EthereumTxReceiptLog,
+        tx_log: EvmTxReceiptLog,
         transaction: EvmTransaction,
         decoded_events: List[HistoryBaseEntry],
-        all_logs: List[EthereumTxReceiptLog],  # pylint: disable=unused-argument
+        all_logs: List[EvmTxReceiptLog],  # pylint: disable=unused-argument
         action_items: Optional[List[ActionItem]],  # pylint: disable=unused-argument
     ) -> Tuple[Optional[HistoryBaseEntry], List[ActionItem]]:
         if tx_log.topics[0] in (
@@ -208,7 +208,7 @@ class CurveDecoder(DecoderInterface):  # lgtm[py/missing-call-to-init]
     @staticmethod
     def _maybe_enrich_curve_transfers(
             token: EvmToken,  # pylint: disable=unused-argument
-            tx_log: EthereumTxReceiptLog,
+            tx_log: EvmTxReceiptLog,
             transaction: EvmTransaction,
             event: HistoryBaseEntry,
             action_items: List[ActionItem],  # pylint: disable=unused-argument
